fix: moved pre-filter from prepare_and_solve to run_solver
